# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "下载管理系统 API",
    "description": "下载管理系统后端接口文档",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/api",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "用户注册",
        "description": "创建新用户账户",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或用户已存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "用户登录",
        "description": "用户登录获取访问令牌",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "summary": "获取用户信息",
        "description": "获取当前登录用户的个人信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/downloads": {
      "get": {
        "summary": "获取下载列表",
        "description": "获取当前用户的下载任务列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Download" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "创建下载任务",
        "description": "创建新的下载任务",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Download" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/downloads/{id}": {
      "get": {
        "summary": "获取下载任务详情",
        "description": "获取指定ID的下载任务详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "下载任务ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Download" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "下载任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新下载任务",
        "description": "更新下载任务状态和进度",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "下载任务ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/Download" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "下载任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除下载任务",
        "description": "删除指定的下载任务",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "下载任务ID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "下载任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/uploads": {
      "get": {
        "summary": "获取上传列表",
        "description": "获取当前用户的上传文件列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Upload" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["username", "email", "password"],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱地址"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "密码"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["password"],
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名（与email二选一）"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "邮箱地址（与username二选一）"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "密码"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "token": { "type": "string" },
              "user": { "$ref": "#/components/schemas/User" }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "DownloadRequest": {
        "type": "object",
        "required": ["url", "filename"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "下载URL"
          },
          "filename": {
            "type": "string",
            "description": "文件名"
          }
        }
      },
      "DownloadUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["pending", "downloading", "completed", "error", "cancelled"],
            "description": "下载状态"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "description": "下载进度"
          },
          "downloaded_bytes": {
            "type": "integer",
            "description": "已下载字节数"
          },
          "total_bytes": {
            "type": "integer",
            "description": "总字节数"
          },
          "speed": {
            "type": "integer",
            "description": "下载速度（字节/秒）"
          },
          "resume_position": {
            "type": "integer",
            "description": "断点续传位置"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "description": "完成时间"
          }
        }
      },
      "Download": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "url": { "type": "string" },
          "filename": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "downloading", "completed", "error", "cancelled"]
          },
          "progress": { "type": "number" },
          "downloaded_bytes": { "type": "integer" },
          "total_bytes": { "type": "integer" },
          "speed": { "type": "integer" },
          "resume_position": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "Upload": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "filename": { "type": "string" },
          "original_filename": { "type": "string" },
          "file_path": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["pending", "uploading", "completed", "error", "cancelled"]
          },
          "progress": { "type": "number" },
          "uploaded_bytes": { "type": "integer" },
          "total_bytes": { "type": "integer" },
          "speed": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "error": { "type": "string", "nullable": true },
          "timestamp": { "type": "string", "format": "date-time" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT认证令牌"
      }
    }
  }
}